update sidemenu routing
